[refactor] dto 리팩토링 (#101)
